backend/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
dotenv==0.9.9
executing==2.2.0
fastjsonschema==2.21.1
idna==3.10
importlib_metadata==8.6.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nulltype==2.3.1
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
plaid-python==29.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.21.0
